updated table's search coverage
